+++ src/app/dapp/_components/mining.tsx
@@ -1,5 +1,4 @@
 import { Card, CardContent } from "@/components/ui/card";
-import { Heading } from "@/components/ui/heading";
 import { Separator } from "@/components/ui/separator";
 import { Balance } from "@/modules/mining/components/balance";
 import { Form } from "@/modules/mining/components/form";
@@ -12,11 +11,8 @@ export function Mining() {
   return (
     <section
       id="mining"
-      className="bg-primary text-primary-foreground space-y-5 px-4 py-16"
+      className="bg-primary text-primary-foreground px-4 py-16"
     >
-      <Heading variant="h2" className="text-center uppercase">
-        Mining
-      </Heading>
       <div className="mx-auto grid w-full max-w-6xl grid-cols-1 items-start gap-6 lg:grid-cols-2">
         <Card className="backdrop-blur-sm">
           <CardContent>
