+++ src/modules/landing/components/hero.tsx
@@ -1,54 +1,63 @@
 import Image from "next/image";
-import Link from "next/link";
 
-import penguins from "@/assets/images/cute-pudgy-penguins-on-ice-floe-with-snowflakes-ar.png";
-import { Box } from "@/components/ui/box";
-import { Button } from "@/components/ui/button";
+import penguin1 from "@/assets/images/penguin_1.png";
+import penguin2 from "@/assets/images/penguin_2.png";
+import penguin3 from "@/assets/images/penguin_3.png";
+import penguin4 from "@/assets/images/penguin_4.png";
+import penguin5 from "@/assets/images/penguin_5.png";
+import snowMountains from "@/assets/images/snow-mountains.png";
+import { Paragraph } from "@/components/ui/paragraph";
+import { Title } from "./title";
 
 export function Hero() {
   return (
     <section
       id="home"
-      className="from-background to-muted relative overflow-hidden bg-gradient-to-b py-20 lg:py-32"
+      className="text-primary-foreground relative h-svh overflow-hidden bg-[#2b90cc]"
     >
-      <Box variant="container">
-        <div className="grid items-center gap-12 lg:grid-cols-2">
-          <div className="space-y-8">
-            <h1
-              className="text-foreground text-4xl leading-tight font-black lg:text-6xl"
-              style={{ fontFamily: "var(--font-heading)" }}
-            >
-              Dive into the World of{" "}
-              <span className="text-primary">Pudgy Penguins!</span>
-            </h1>
-
-            <p className="text-muted-foreground max-w-lg text-lg leading-relaxed">
-              Join the coolest penguin community in the NFT space. Discover
-              unique, adorable penguins and become part of our arctic adventure
-              filled with exclusive benefits and endless fun.
-            </p>
-
-            <div className="flex flex-col gap-4 sm:flex-row">
-              <Button size="lg">
-                <Link href="/dapp">Join the Adventure</Link>
-              </Button>
-              <Button variant="outline" size="lg">
-                Explore Collection
-              </Button>
-            </div>
-          </div>
-
-          <div className="relative">
-            <Image
-              src={penguins}
-              alt="Pudgy Penguins on Ice"
-              className="h-auto w-full rounded-2xl shadow-2xl"
-            />
-            <div className="bg-accent/20 absolute -top-4 -right-4 h-20 w-20 rounded-full blur-xl" />
-            <div className="bg-primary/20 absolute -bottom-4 -left-4 h-32 w-32 rounded-full blur-xl" />
-          </div>
+      <Image
+        src={snowMountains}
+        alt="Snow mountains background"
+        priority
+        className="absolute inset-x-0 bottom-0 h-full w-full object-cover"
+      />
+      <div className="relative z-1 flex h-full flex-col">
+        <div className="absolute inset-0 bg-black/20" />
+        <div className="relative z-1 flex flex-col items-center bg-linear-180 from-[#2b90cc] from-70% to-transparent px-[10svw] pt-[10svh]">
+          <Title />
+          <Paragraph className="mt-10 max-w-2xl text-center text-xl font-black uppercase">
+            Pudgy Penguins is a global IP focused on proliferating the penguin,
+            memetic culture, and good vibes.
+          </Paragraph>
         </div>
-      </Box>
+        <div className="relative mx-auto hidden h-full w-full lg:block">
+          <Image
+            src={penguin1}
+            alt="Penguin 1"
+            className="absolute bottom-0 left-0 z-2 h-auto max-h-72 w-full object-contain"
+          />
+          <Image
+            src={penguin2}
+            alt="Penguin 2"
+            className="absolute bottom-0 left-0 z-1 h-auto max-h-80 w-full object-contain"
+          />
+          <Image
+            src={penguin3}
+            alt="Penguin 3"
+            className="absolute bottom-0 left-0 z-2 h-auto max-h-96 w-full object-contain"
+          />
+          <Image
+            src={penguin4}
+            alt="Penguin 4"
+            className="absolute bottom-0 left-0 z-1 h-auto max-h-80 w-full object-contain"
+          />
+          <Image
+            src={penguin5}
+            alt="Penguin 5"
+            className="absolute bottom-0 left-0 z-2 h-auto max-h-72 w-full object-contain"
+          />
+        </div>
+      </div>
     </section>
   );
 }
